plan: extract getChildrenPossibleProp to a file. (#5286)
