+++ plan/physical_plan_builder.go
@@ -19,7 +19,6 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/juju/errors"
 	"github.com/pingcap/tidb/ast"
-	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/kv"
@@ -49,197 +48,6 @@ var JoinConcurrency = 5
 var wholeTaskTypes = [...]taskType{copSingleReadTaskType, copDoubleReadTaskType, rootTaskType}
 
 var invalidTask = &rootTask{cst: math.MaxFloat64}
-
-func (p *requiredProp) enforceProperty(task task, ctx context.Context) task {
-	if p.isEmpty() {
-		return task
-	}
-	// If task is invalid, keep it remained.
-	if task.plan() == nil {
-		return task
-	}
-	task = finishCopTask(task, ctx)
-	sort := Sort{ByItems: make([]*ByItems, 0, len(p.cols))}.init(ctx)
-	for _, col := range p.cols {
-		sort.ByItems = append(sort.ByItems, &ByItems{col, p.desc})
-	}
-	sort.SetSchema(task.plan().Schema())
-	sort.profile = task.plan().statsProfile()
-	return sort.attach2Task(task)
-}
-
-func (p *PhysicalUnionScan) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	return [][]*requiredProp{{prop}}
-}
-
-// getChildrenPossibleProps will check if this sort property can be pushed or not.
-// When a sort column will be replaced by scalar function, we refuse it.
-// When a sort column will be replaced by a constant, we just remove it.
-func (p *Projection) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	newProp := &requiredProp{taskTp: rootTaskType, expectedCnt: prop.expectedCnt}
-	newCols := make([]*expression.Column, 0, len(prop.cols))
-	for _, col := range prop.cols {
-		idx := p.schema.ColumnIndex(col)
-		if idx == -1 {
-			return nil
-		}
-		switch expr := p.Exprs[idx].(type) {
-		case *expression.Column:
-			newCols = append(newCols, expr)
-		case *expression.ScalarFunction:
-			return nil
-		}
-	}
-	newProp.cols = newCols
-	newProp.desc = prop.desc
-	return [][]*requiredProp{{newProp}}
-}
-
-// joinKeysMatchIndex checks if all keys match columns in index.
-func joinKeysMatchIndex(keys []*expression.Column, index *model.IndexInfo) []int {
-	if len(index.Columns) < len(keys) {
-		return nil
-	}
-	matchOffsets := make([]int, len(keys))
-	for i, idxCol := range index.Columns {
-		if idxCol.Length != types.UnspecifiedLength {
-			return nil
-		}
-		found := false
-		for j, key := range keys {
-			if idxCol.Name.L == key.ColName.L {
-				matchOffsets[i] = j
-				found = true
-				break
-			}
-		}
-		if !found {
-			return nil
-		}
-		if i+1 == len(keys) {
-			break
-		}
-	}
-	return matchOffsets
-}
-
-func (p *LogicalJoin) constructIndexJoin(innerJoinKeys, outerJoinKeys []*expression.Column, outerIdx int, innerPlan PhysicalPlan) []PhysicalPlan {
-	joinType := p.JoinType
-	join := PhysicalIndexJoin{
-		OuterIndex:      outerIdx,
-		LeftConditions:  p.LeftConditions,
-		RightConditions: p.RightConditions,
-		OtherConditions: p.OtherConditions,
-		JoinType:        joinType,
-		OuterJoinKeys:   outerJoinKeys,
-		InnerJoinKeys:   innerJoinKeys,
-		DefaultValues:   p.DefaultValues,
-		outerSchema:     p.children[outerIdx].Schema(),
-		innerPlan:       innerPlan,
-	}.init(p.ctx, p.children...)
-	join.SetSchema(p.schema)
-	join.profile = p.profile
-	orderJoin := join.Copy().(*PhysicalIndexJoin)
-	orderJoin.KeepOrder = true
-	return []PhysicalPlan{join, orderJoin}
-}
-
-// getIndexJoinByOuterIdx will generate index join by outerIndex. OuterIdx points out the outer child,
-// because we will swap the children of join when the right child is outer child.
-// First of all, we will extract the join keys for p's equal conditions. If the join keys can match some of the indices or PK
-// column of inner child, we can apply the index join.
-func (p *LogicalJoin) getIndexJoinByOuterIdx(outerIdx int) []PhysicalPlan {
-	innerChild := p.children[1-outerIdx].(LogicalPlan)
-	var (
-		usedIndexInfo *model.IndexInfo
-		innerJoinKeys []*expression.Column
-		outerJoinKeys []*expression.Column
-	)
-	if outerIdx == 0 {
-		outerJoinKeys = p.LeftJoinKeys
-		innerJoinKeys = p.RightJoinKeys
-	} else {
-		innerJoinKeys = p.LeftJoinKeys
-		outerJoinKeys = p.RightJoinKeys
-	}
-	x, ok := innerChild.(*DataSource)
-	if !ok {
-		return nil
-	}
-	indices := x.availableIndices.indices
-	includeTableScan := x.availableIndices.includeTableScan
-	if includeTableScan && len(innerJoinKeys) == 1 {
-		pkCol := x.getPKIsHandleCol()
-		if pkCol != nil && innerJoinKeys[0].Equal(pkCol, nil) {
-			innerPlan := x.forceToTableScan()
-			return p.constructIndexJoin(innerJoinKeys, outerJoinKeys, outerIdx, innerPlan)
-		}
-	}
-	for _, indexInfo := range indices {
-		matchedOffsets := joinKeysMatchIndex(innerJoinKeys, indexInfo)
-		if matchedOffsets == nil {
-			continue
-		}
-		usedIndexInfo = indexInfo
-		newOuterJoinKeys := make([]*expression.Column, len(outerJoinKeys))
-		newInnerJoinKeys := make([]*expression.Column, len(innerJoinKeys))
-		for i, offset := range matchedOffsets {
-			newOuterJoinKeys[i] = outerJoinKeys[offset]
-			newInnerJoinKeys[i] = innerJoinKeys[offset]
-		}
-		outerJoinKeys = newOuterJoinKeys
-		innerJoinKeys = newInnerJoinKeys
-		break
-	}
-	if usedIndexInfo != nil {
-		innerPlan := x.forceToIndexScan(usedIndexInfo)
-		return p.constructIndexJoin(innerJoinKeys, outerJoinKeys, outerIdx, innerPlan)
-	}
-	return nil
-}
-
-// getChildrenPossibleProps gets children possible props.:
-// For index join, we shouldn't require a root task which may let CBO framework select a sort operator in fact.
-// We are not sure which way of index scanning we should choose, so we try both single read and double read and finally
-// it will result in a best one.
-func (p *PhysicalIndexJoin) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	if !prop.isEmpty() && !p.KeepOrder {
-		return nil
-	}
-	for _, col := range prop.cols {
-		if p.outerSchema.ColumnIndex(col) == -1 {
-			return nil
-		}
-	}
-	requiredProps1 := make([]*requiredProp, 2)
-	requiredProps1[p.OuterIndex] = &requiredProp{taskTp: rootTaskType, expectedCnt: prop.expectedCnt, cols: prop.cols, desc: prop.desc}
-	requiredProps1[1-p.OuterIndex] = &requiredProp{taskTp: rootTaskType, expectedCnt: math.MaxFloat64}
-	return [][]*requiredProp{requiredProps1}
-}
-
-func (p *PhysicalMergeJoin) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	lProp := &requiredProp{taskTp: rootTaskType, cols: p.leftKeys, expectedCnt: math.MaxFloat64}
-	rProp := &requiredProp{taskTp: rootTaskType, cols: p.rightKeys, expectedCnt: math.MaxFloat64}
-	if !prop.isEmpty() {
-		if prop.desc {
-			return nil
-		}
-		if !prop.isPrefix(lProp) && !prop.isPrefix(rProp) {
-			return nil
-		}
-		if prop.isPrefix(rProp) && p.JoinType == LeftOuterJoin {
-			return nil
-		}
-		if prop.isPrefix(lProp) && p.JoinType == RightOuterJoin {
-			return nil
-		}
-	}
-
-	return [][]*requiredProp{{lProp, rProp}}
-}
 
 // getPropByOrderByItems will check if this sort property can be pushed or not. In order to simplify the problem, we only
 // consider the case that all expression are columns and all of them are asc or desc.
@@ -856,106 +664,4 @@ func (ts *PhysicalTableScan) addPushedDownSelection(copTask *copTask, profile *s
 		// FIXME: It seems wrong...
 		copTask.cst += copTask.count() * cpuFactor
 	}
-}
-
-func (p *basePhysicalPlan) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.basePlan.expectedCnt = prop.expectedCnt
-	// By default, physicalPlan can always match the orders.
-	props := make([]*requiredProp, 0, len(p.basePlan.children))
-	for range p.basePlan.children {
-		props = append(props, prop)
-	}
-	return [][]*requiredProp{props}
-}
-
-func (p *PhysicalSelection) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	return [][]*requiredProp{{prop}}
-}
-
-func (p *PhysicalHashJoin) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	if !prop.isEmpty() {
-		return nil
-	}
-	return [][]*requiredProp{{&requiredProp{taskTp: rootTaskType, expectedCnt: prop.expectedCnt}, &requiredProp{taskTp: rootTaskType, expectedCnt: math.MaxFloat64}}}
-}
-
-func (p *PhysicalHashSemiJoin) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	lProp := &requiredProp{taskTp: rootTaskType, cols: prop.cols, expectedCnt: prop.expectedCnt, desc: prop.desc}
-	for _, col := range lProp.cols {
-		idx := p.Schema().ColumnIndex(col)
-		if idx == -1 || idx >= p.rightChOffset {
-			return nil
-		}
-	}
-	return [][]*requiredProp{{lProp, &requiredProp{taskTp: rootTaskType, expectedCnt: math.MaxFloat64}}}
-}
-
-func (p *PhysicalApply) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	lProp := &requiredProp{taskTp: rootTaskType, cols: prop.cols, expectedCnt: prop.expectedCnt, desc: prop.desc}
-	for _, col := range lProp.cols {
-		idx := p.Schema().ColumnIndex(col)
-		if idx == -1 || idx >= p.rightChOffset {
-			return nil
-		}
-	}
-	return [][]*requiredProp{{lProp, &requiredProp{taskTp: rootTaskType, expectedCnt: math.MaxFloat64}}}
-}
-
-func (p *Limit) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	if !prop.isEmpty() {
-		return nil
-	}
-	props := make([][]*requiredProp, 0, len(wholeTaskTypes))
-	for _, tp := range wholeTaskTypes {
-		newProp := &requiredProp{taskTp: tp, expectedCnt: float64(p.Count + p.Offset)}
-		if p.expectedProp != nil {
-			newProp.cols = p.expectedProp.cols
-			newProp.desc = p.expectedProp.desc
-		}
-		props = append(props, []*requiredProp{newProp})
-	}
-	return props
-}
-
-func (p *TopN) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	if !prop.isEmpty() {
-		return nil
-	}
-	props := make([][]*requiredProp, 0, len(wholeTaskTypes))
-	for _, tp := range wholeTaskTypes {
-		props = append(props, []*requiredProp{{taskTp: tp, expectedCnt: math.MaxFloat64}})
-	}
-	return props
-}
-
-func (p *PhysicalHashAgg) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	if !prop.isEmpty() {
-		return nil
-	}
-	props := make([][]*requiredProp, 0, len(wholeTaskTypes))
-	for _, tp := range wholeTaskTypes {
-		props = append(props, []*requiredProp{{taskTp: tp, expectedCnt: math.MaxFloat64}})
-	}
-	return props
-}
-
-func (p *PhysicalStreamAgg) getChildrenPossibleProps(prop *requiredProp) [][]*requiredProp {
-	p.expectedCnt = prop.expectedCnt
-	reqProp := &requiredProp{
-		taskTp:      rootTaskType,
-		cols:        p.propKeys,
-		expectedCnt: prop.expectedCnt * p.inputCount / p.profile.count,
-		desc:        prop.desc,
-	}
-	if !prop.isEmpty() && !prop.isPrefix(reqProp) {
-		return nil
-	}
-	return [][]*requiredProp{{reqProp}}
 }
